avoid of proccessing existed tweet - clean coding
