+++ auto-kol/src/services/agents/workflow.ts
@@ -1,26 +1,17 @@
 import { END, MemorySaver, StateGraph, START, Annotation } from '@langchain/langgraph';
-import { AIMessage, BaseMessage, SystemMessage } from '@langchain/core/messages';
+import { AIMessage, BaseMessage } from '@langchain/core/messages';
 import { ChatOpenAI } from '@langchain/openai';
-import { ChatPromptTemplate, PromptTemplate } from '@langchain/core/prompts';
 import { MessageContent } from '@langchain/core/messages';
 import { config } from '../../config/index.js';
 import { createLogger } from '../../utils/logger.js';
 import { createTools } from '../../tools/index.js';
 import { createTwitterClient } from '../twitter/api.js';
-import {
-    engagementSchema,
-    tweetSearchSchema,
-    toneSchema,
-    responseSchema
-} from '../../schemas/workflow.js';
 import { ToolNode } from '@langchain/langgraph/prebuilt';
 import { TwitterApiReadWrite } from 'twitter-api-v2';
 import { StructuredOutputParser } from 'langchain/output_parsers';
-import { ChromaService } from '../vectorstore/chroma.js';
-
-const logger = createLogger('agent-workflow');
-
-const parseMessageContent = (content: MessageContent): any => {
+export  const logger = createLogger('agent-workflow');
+import { createNodes } from './nodes.js';
+export const parseMessageContent = (content: MessageContent): any => {
     if (typeof content === 'string') {
         return JSON.parse(content);
     }
@@ -31,7 +22,7 @@ const parseMessageContent = (content: MessageContent): any => {
 };
 
 // State management with processed tweets tracking
-const State = Annotation.Root({
+export const State = Annotation.Root({
     messages: Annotation<readonly BaseMessage[]>({
         reducer: (curr, prev) => [...curr, ...prev],
         default: () => [],
@@ -44,7 +35,7 @@ const State = Annotation.Root({
 });
 
 // Workflow configuration type
-type WorkflowConfig = Readonly<{
+export type WorkflowConfig = Readonly<{
     client: TwitterApiReadWrite;
     toolNode: ToolNode;
     llms: Readonly<{
@@ -88,420 +79,6 @@ const createWorkflowConfig = async (): Promise<WorkflowConfig> => {
     };
 };
 
-// Node factory functions
-const createNodes = async (config: WorkflowConfig) => {
-
-    // Placeholder for now, because the timeline tool doesn't work
-    const timelineNode = async (state: typeof State.State) => {
-        logger.info('Timeline Node - Fetching recent tweets');
-        const toolResponse = await config.toolNode.invoke({
-            messages: [
-                new AIMessage({
-                    content: '',
-                    tool_calls: [{
-                        name: 'fetch_timeline',
-                        args: {},
-                        id: 'fetch_timeline_call',
-                        type: 'tool_call'
-                    }]
-                })
-            ]
-        });
-
-        // Log the tool response
-        logger.info('Tool response received:', {
-            messageCount: toolResponse.messages.length,
-            lastMessageContent: toolResponse.messages[toolResponse.messages.length - 1].content
-        });
-
-        // Parsing the ALL the tweets
-        const parsedTweets = tweetSearchSchema.parse(toolResponse.messages[toolResponse.messages.length - 1].content);
-
-        // TODO: Check if the tweets are new | doesn't exist in sqlite db
-
-        const chromaService = await ChromaService.getInstance();
-        for (const tweet of parsedTweets.tweets) {
-            await chromaService.addTweet(tweet);
-        }
-
-        logger.info(`Added ${parsedTweets.tweets.length} tweets to the vector db`);
-
-
-        return {
-            messages: [new AIMessage({
-                content: JSON.stringify(parsedTweets)
-            })],
-            lastProcessedId: parsedTweets.lastProcessedId || undefined
-        };
-    }
-
-    const searchNode = async (state: typeof State.State) => {
-        logger.info('Search Node - Fetching recent tweets');
-        try {
-            logger.info('Last processed id:', state.lastProcessedId);
-
-            const toolResponse = await config.toolNode.invoke({
-                messages: [
-                    new AIMessage({
-                        content: '',
-                        tool_calls: [{
-                            name: 'search_recent_tweets',
-                            args: {
-                                lastProcessedId: state.lastProcessedId || undefined
-                            },
-                            id: 'tool_call_id',
-                            type: 'tool_call'
-                        }],
-                    }),
-                ],
-            });
-
-            // Log the tool response
-            logger.info('Tool response received:', {
-                messageCount: toolResponse.messages.length,
-                lastMessageContent: toolResponse.messages[toolResponse.messages.length - 1].content
-            });
-
-            // The tool response will be in the last message's content
-            const lastMessage = toolResponse.messages[toolResponse.messages.length - 1];
-
-            // Handle different types of content
-            let searchResult;
-            if (typeof lastMessage.content === 'string') {
-                try {
-                    searchResult = JSON.parse(lastMessage.content);
-                    logger.info('Parsed search result:', searchResult);
-                } catch (error) {
-                    logger.error('Failed to parse search result:', error);
-                    searchResult = { tweets: [], lastProcessedId: null };
-                }
-            } else {
-                searchResult = lastMessage.content;
-                logger.info('Non-string search result:', searchResult);
-            }
-
-            // Validate the search result
-            const validatedResult = tweetSearchSchema.parse(searchResult);
-
-
-            // TODO: Check if the tweets are new | doesn't exist in sqlite db
-            if (validatedResult.tweets.length > 0) {
-                const chromaService = await ChromaService.getInstance();
-                for (const tweet of validatedResult.tweets) {
-                    await chromaService.addTweet(tweet);
-                }
-            }
-
-            logger.info(`Found ${validatedResult.tweets.length} tweets`);
-
-            return {
-                messages: [new AIMessage({
-                    content: JSON.stringify({
-                        tweets: validatedResult.tweets,
-                        currentTweetIndex: 0,
-                        lastProcessedId: validatedResult.lastProcessedId
-                    })
-                })],
-                lastProcessedId: validatedResult.lastProcessedId || undefined
-            };
-        } catch (error) {
-            logger.error('Error in search node:', error);
-            // Return a valid empty result
-            const emptyResult = {
-                tweets: [],
-                lastProcessedId: null
-            };
-            return {
-                messages: [new AIMessage({ content: JSON.stringify(emptyResult) })],
-                lastProcessedId: undefined
-            };
-        }
-    };
-
-    const engagementParser = StructuredOutputParser.fromZodSchema(engagementSchema);
-    const toneParser = StructuredOutputParser.fromZodSchema(toneSchema);
-    const responseParser = StructuredOutputParser.fromZodSchema(responseSchema);
-
-    const engagementSystemPrompt = await PromptTemplate.fromTemplate(
-        "You are a strategic social media engagement advisor. Your task is to evaluate tweets and decide whether they warrant a response.\n\n" +
-        "Evaluate each tweet based on:\n" +
-        "1. Relevance to AI, blockchain, or tech innovation\n" +
-        "2. Potential for meaningful discussion, entertainment, or argumentation\n" +
-        "3. Author's influence and engagement level\n" +
-        "4. Tweet's recency and context\n\n" +
-        "{format_instructions}"
-    ).format({
-        format_instructions: engagementParser.getFormatInstructions()
-    });
-
-    const toneSystemPrompt = await PromptTemplate.fromTemplate(
-        "You are an expert in social media tone analysis. Your task is to analyze the tone of tweets and suggest the best tone for responses.\n\n" +
-        "Feel free to be argumentative. Consider:\n" +
-        "1. The original tweet's tone and context\n" +
-        "2. The author's typical communication style\n" +
-        "3. The topic and its sensitivity\n" +
-        "4. The platform's culture\n\n" +
-        "{format_instructions}"
-    ).format({
-        format_instructions: toneParser.getFormatInstructions()
-    });
-
-    const responseSystemPrompt = await PromptTemplate.fromTemplate(
-        "You are a sharp-witted crypto & AI expert known for your direct, often contrarian takes on Twitter. Your responses are naturally argumentative and carry a hint of sass, but always grounded in technical knowledge.\n\n" +
-        "Core Traits:\n" +
-        "- ALWAYS reply to the author of the tweet, not any users mentioned in the tweet\n" +
-        "- Confident in your technical expertise\n" +
-        "- Naturally argumentative without being formulaic\n" +
-        "- Quick to point out technical flaws\n" +
-        "- Slightly cold but intellectually engaging\n" +
-        "- More interested in technical truth than being agreeable\n\n" +
-        "Response Philosophy:\n" +
-        "- Form your own technical analysis first\n" +
-        "- Challenge assumptions naturally\n" +
-        "- Let your personality shine through varied expressions\n" +
-        "- Historical context should complement, not drive your response\n" +
-        "- Each response should feel fresh and specific to the discussion\n\n" +
-        "Remember: You're having a real technical discussion, not following a formula. Your expertise and slightly sassy personality should come through naturally in each unique response.\n\n" +
-        "{format_instructions}"
-    ).format({
-        format_instructions: responseParser.getFormatInstructions()
-    });
-
-    const engagementPrompt = ChatPromptTemplate.fromMessages([
-        new SystemMessage(engagementSystemPrompt),
-        ["human", "Evaluate this tweet and provide your structured decision: {tweet}"]
-    ]);
-
-    const tonePrompt = ChatPromptTemplate.fromMessages([
-        new SystemMessage(toneSystemPrompt),
-        ["human", "Analyze the tone for this tweet and suggest a response tone: {tweet}"]
-    ]);
-
-    const responsePrompt = ChatPromptTemplate.fromMessages([
-        new SystemMessage(responseSystemPrompt),
-        ["human", `Generate a response strategy for this tweet by considering similar tweets from @{author} using the suggested tone:
-        Tweet: {tweet}
-        Tone: {tone}
-        Similar Tweets: {similarTweets}
-
-        Instructions:
-        1. Review the similar tweets to identify any inconsistencies or changes in the author's stance
-        2. If you find contradictions, incorporate them into your response
-        3. Use phrases like "Interesting shift from your previous stance where..." or "This seems to contradict your earlier view that..."
-        4. Be specific when referencing past statements
-        `]
-    ]);
-
-    const engagementNode = async (state: typeof State.State) => {
-        logger.info('Engagement Node - Evaluating tweet engagement');
-        try {
-            const lastMessage = state.messages[state.messages.length - 1];
-            const parsedContent = parseMessageContent(lastMessage.content);
-
-            if (!parsedContent.tweets ||
-                parsedContent.currentTweetIndex >= parsedContent.tweets.length) {
-                logger.info('No more tweets to process');
-                return { messages: [] };
-            }
-
-            const tweet = parsedContent.tweets[parsedContent.currentTweetIndex];
-
-            if (state.processedTweets.has(tweet.id)) {
-                logger.info(`Tweet ${tweet.id} already processed, moving to next`);
-                return {
-                    messages: [new AIMessage({
-                        content: JSON.stringify({
-                            tweets: parsedContent.tweets,
-                            currentTweetIndex: parsedContent.currentTweetIndex + 1,
-                            lastProcessedId: parsedContent.lastProcessedId
-                        })
-                    })]
-                };
-            }
-
-            const decision = await engagementPrompt
-                .pipe(config.llms.decision)
-                .pipe(engagementParser)
-                .invoke({ tweet: tweet.text });
-
-            logger.info('LLM decision:', { decision });
-
-            if (!decision.shouldEngage) {
-                logger.info('Queueing skipped tweet to review queue:', { tweetId: tweet.id });
-
-                // Ensure tweet object matches schema
-                const tweetData = {
-                    id: tweet.id,
-                    text: tweet.text,
-                    authorId: tweet.authorId,
-                    authorUsername: tweet.authorUsername || 'unknown', // Ensure we have a fallback
-                    createdAt: typeof tweet.createdAt === 'string' ? tweet.createdAt : tweet.createdAt.toISOString()
-                };
-
-                const toolResult = await config.toolNode.invoke({
-                    messages: [new AIMessage({
-                        content: '',
-                        tool_calls: [{
-                            name: 'queue_skipped',
-                            args: {
-                                tweet: tweetData,
-                                reason: decision.reason,
-                                priority: decision.priority || 0,
-                                workflowState: { decision }
-                            },
-                            id: 'skip_call',
-                            type: 'tool_call'
-                        }]
-                    })]
-                });
-
-                logger.info('Queue skipped result:', toolResult);
-
-                return {
-                    messages: [new AIMessage({
-                        content: JSON.stringify({
-                            tweets: parsedContent.tweets,
-                            currentTweetIndex: parsedContent.currentTweetIndex + 1,
-                            lastProcessedId: parsedContent.lastProcessedId
-                        })
-                    })],
-                    processedTweets: new Set([tweet.id])
-                };
-            }
-
-            return {
-                messages: [new AIMessage({
-                    content: JSON.stringify({
-                        tweets: parsedContent.tweets,
-                        currentTweetIndex: parsedContent.currentTweetIndex,
-                        tweet,
-                        decision
-                    })
-                })]
-            };
-        } catch (error) {
-            logger.error('Error in engagement node:', error);
-            return { messages: [] };
-        }
-    };
-
-    const toneAnalysisNode = async (state: typeof State.State) => {
-        logger.info('Tone Analysis Node - Analyzing tweet tone');
-        try {
-            const lastMessage = state.messages[state.messages.length - 1];
-            const parsedContent = parseMessageContent(lastMessage.content);
-            const { tweet, tweets, currentTweetIndex } = parsedContent;
-
-            const toneAnalysis = await tonePrompt
-                .pipe(config.llms.tone)
-                .pipe(toneParser)
-                .invoke({ tweet: tweet.text });
-
-            logger.info('Tone analysis:', { toneAnalysis });
-
-            return {
-                messages: [new AIMessage({
-                    content: JSON.stringify({
-                        tweets,
-                        currentTweetIndex,
-                        tweet,
-                        toneAnalysis
-                    })
-                })]
-            };
-        } catch (error) {
-            logger.error('Error in tone analysis node:', error);
-            return { messages: [] };
-        }
-    };
-
-    const responseGenerationNode = async (state: typeof State.State) => {
-        logger.info('Response Generation Node - Creating response strategy');
-        try {
-            const lastMessage = state.messages[state.messages.length - 1];
-            const parsedContent = parseMessageContent(lastMessage.content);
-            const { tweet, toneAnalysis, tweets, currentTweetIndex } = parsedContent;
-
-            // Search for similar tweets by this author
-            const similarTweetsResponse = await config.toolNode.invoke({
-                messages: [new AIMessage({
-                    content: '',
-                    tool_calls: [{
-                        name: 'search_similar_tweets',
-                        args: {
-                            query: `author:${tweet.authorUsername} ${tweet.text}`,
-                            k: 5
-                        },
-                        id: 'similar_tweets_call',
-                        type: 'tool_call'
-                    }],
-                })],
-            });
-
-            // Parse similar tweets response
-            const similarTweets = parseMessageContent(
-                similarTweetsResponse.messages[similarTweetsResponse.messages.length - 1].content
-            );
-
-            // Include similar tweets in the response prompt context
-            const responseStrategy = await responsePrompt
-                .pipe(config.llms.response)
-                .pipe(responseParser)
-                .invoke({
-                    tweet: tweet.text,
-                    tone: toneAnalysis.suggestedTone,
-                    author: tweet.authorUsername,
-                    similarTweets: JSON.stringify(similarTweets.similar_tweets)
-                });
-
-            logger.info('Response strategy:', { responseStrategy });
-
-            // Queue the response
-            await config.toolNode.invoke({
-                messages: [new AIMessage({
-                    content: '',
-                    tool_calls: [{
-                        name: 'queue_response',
-                        args: {
-                            tweet,
-                            response: responseStrategy.content,
-                            workflowState: {
-                                toneAnalysis,
-                                responseStrategy
-                            }
-                        },
-                        id: 'queue_response_call',
-                        type: 'tool_call'
-                    }]
-                })]
-            });
-
-            // Move to next tweet and add current to processed set
-            return {
-                messages: [new AIMessage({
-                    content: JSON.stringify({
-                        tweets,
-                        currentTweetIndex: currentTweetIndex + 1,
-                        lastProcessedId: tweet.id,
-                        responseStrategy
-                    })
-                })],
-                processedTweets: new Set([tweet.id])
-            };
-        } catch (error) {
-            logger.error('Error in response generation node:', error);
-            return { messages: [] };
-        }
-    };
-
-    return {
-        timelineNode,
-        searchNode,
-        engagementNode,
-        toneAnalysisNode,
-        responseGenerationNode
-    };
-};
 
 const shouldContinue = (state: typeof State.State) => {
     const lastMessage = state.messages[state.messages.length - 1];
